test of redirecting project location.
I added to the fact book array to see how changes made in the program's
working directory are reflected in the building of the android app.  It
turns out that it is a little more complicated than copying and pasting
the project files from one directory to the other.  You have to sync
gradle with the project files.